Trait stc_ts_file_analyzer::loader::Load

pub trait Load: 'static + Send + Sync {
    fn module_id(&self, base: &Arc<FileName>, src: &JsWord) -> Option<ModuleId>;
    fn is_in_same_circular_group(&self, base: ModuleId, dep: ModuleId) -> bool;
    fn load_circular_dep(
        &self,
        base: ModuleId,
        dep: ModuleId,
        partial: &ModuleTypeData
    ) -> VResult<Type>;
    fn load_non_circular_dep(
        &self,
        base: ModuleId,
        dep: ModuleId
    ) -> VResult<Type>;
    fn declare_module(&self, name: &JsWord, module: Type);
}

Group of circular imports are handled by one thread. This

Required Methods

fn module_id(&self, base: &Arc<FileName>, src: &JsWord) -> Option<ModuleId>

fn is_in_same_circular_group(&self, base: ModuleId, dep: ModuleId) -> bool

Note: This method called within a thread

fn load_circular_dep(
&self,
base: ModuleId,
dep: ModuleId,
partial: &ModuleTypeData
) -> VResult<Type>

This method can be called multiple time for same module.

Also note that this method is called within a single thread.

partial denotes the types and variables which the [Analyzer] successed processing, with resolved imports.

Returned value must be Type::Arc of Type::Module

fn load_non_circular_dep(&self, base: ModuleId, dep: ModuleId) -> VResult<Type>

Note: This method is called in parallel.

Returned value must be Type::Arc of Type::Module

fn declare_module(&self, name: &JsWord, module: Type)

module should be Type::Arc of Type::Module.